Be more systematic about device management.
Create a decorator to pick default device if it's not provided, and run a function in that device context.

In diffusion_utils:modify_conformer, make sure output is the same device as input.

Add a decorator for suppressing repetitive PDB loading warning, and just printing a count at the end.